+++ bigchaindb/backend/__init__.py
@@ -7,7 +7,6 @@ configuration or the ``BIGCHAINDB_DATABASE_BACKEND`` environment variable.
 """
 
 # Include the backend interfaces
-from bigchaindb.backend import admin, changefeed, schema, query  # noqa
+from bigchaindb.backend import schema, query  # noqa
 
 from bigchaindb.backend.connection import connect  # noqa
-from bigchaindb.backend.changefeed import get_changefeed  # noqa
