--- tests/backend/mongodb-ssl/certs/bigchaindb.cert.pem
@@ -1,33 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFnTCCA4WgAwIBAgICEAEwDQYJKoZIhvcNAQELBQAwgasxCzAJBgNVBAYTAlVW
-MRQwEgYDVQQIDAtDb3NtaWNzdGF0ZTEuMCwGA1UECgwlRGVjZW50cmFsaXplZCBO
-b21hZGljIEFudHMgQ29sbGVjdGl2ZTEWMBQGA1UECwwNTGVhZGVyTGVzc0xhYjEc
-MBoGA1UEAwwTSW50ZXJtZWRpYXRlIEROQUNMMzEgMB4GCSqGSIb3DQEJARYRZW1w
-dHlAZG5hYy1sMy5pb24wHhcNMTcxMjIxMDA1NDA4WhcNMTgxMjMxMDA1NDA4WjCB
-rjELMAkGA1UEBhMCVVYxFDASBgNVBAgMC0Nvc21pY3N0YXRlMREwDwYDVQQHDAh6
-ZXJvbmVzdDEuMCwGA1UECgwlRGVjZW50cmFsaXplZCBOb21hZGljIEFudHMgQ29s
-bGVjdGl2ZTEWMBQGA1UECwwNTGVhZGVyTGVzc0xhYjEOMAwGA1UEAwwFYWxpY2Ux
-HjAcBgkqhkiG9w0BCQEWD2FsaWNlQHRoZXJlLnh5ejCCASIwDQYJKoZIhvcNAQEB
-BQADggEPADCCAQoCggEBANflMNZfdb5yySneZxYolgTEV+6kTihIksRiNePJRjfY
-r5eT99S03ds1h/B/Y61r0JkxyadquDj3RhReDY25xMGyZS4M2MyLN2phvR1We7RB
-cxsjFV/qWpwouPwVx0yeeKRxe0z3hCVKD+B51Au1+b2dbS4pPOxrF9tm9VEhtqcq
-cuOpMi8sdAraqGNr3XfzsoHMJ9sqj4bUD8s3SZVfmvnStyIt3liX+fwz5fx94Dc1
-KsySHtcqA5FyEHazXVMNWmDiS6ZINaoIOU44FTAN3pWHg4EbzGUmQN5x2Pqb5WBA
-tBRqUHwhbg2ORcIhT33UODD0azIdf05WuGotMzvuV9MCAwEAAaOBxTCBwjAJBgNV
-HRMEAjAAMBEGCWCGSAGG+EIBAQQEAwIFoDAzBglghkgBhvhCAQ0EJhYkT3BlblNT
-TCBHZW5lcmF0ZWQgQ2xpZW50IENlcnRpZmljYXRlMB0GA1UdDgQWBBQQ8b79Kkf1
-4S6MhAe3oToDV7BHdzAfBgNVHSMEGDAWgBQCF8NWbOOWvQ4kSeE0KVw/SNrTBjAO
-BgNVHQ8BAf8EBAMCBeAwHQYDVR0lBBYwFAYIKwYBBQUHAwIGCCsGAQUFBwMEMA0G
-CSqGSIb3DQEBCwUAA4ICAQCRJoDdgJzsOE8OBWKDaE9KY6Zk+TdnkQKgWQF8VFrh
-lEkprl7RuE3HQKno/u2sk7apz6Hs8fZAH87pQKUl8o/30k3ev+s/oXRKPD3i7lsf
-QUexGEAETTfCdkuXkIVyvq/o5pEXOuEud0i69svP6v7yJ1otb5aUu0e8wTLNpukl
-zICaKzaGuUuoS3vCwMHzBaDpz7ed3kpRS7rUwyMUYQtJ7jpgmGtaw9g4kSXmNyvb
-ABpBJCMtMBkL3gZOfTAELmRSZoZczW1aINtqmLB2ejab1AtyjkuO7iUzJjIG/KQi
-/x2GCSCL4hBe0YSBObZY0uHBq40qQDxVRhcMe36Ch2vc/TPArl/w9tlZR6FhOm3o
-tLsoM1iGj0+oOe9FJA0IAIndP8Hsr1eDJ0tY5IFMY8WsWAh8yLsSYNcdrzv+AYKd
-P+QnSgaIFxvvZ/589WThsGgI2VROZHL+JwvoKyvxdEOMcYFNtzS7sdIcWp2KoIwk
-xi45upt/KuNEbD9+6oTe8adEoa0g93AZn9jIU3MavpgekLwZsiHJyBWz3HkGMo9l
-AZLllLi32dr3i74YCCizRhKfYTC0HAb1Dw0W0s9yl+drtSTf+lzEKegMp4ovwYd8
-29y+y4TNjpq5wlJtUZvFUQXxJZk4FGK8jLIG+HsYKefKB9DXCanSrgW2jdnNTKfS
-Ig==
------END CERTIFICATE-----
