--- tests/backend/mongodb-ssl/certs/bigchaindb.key.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEA1+Uw1l91vnLJKd5nFiiWBMRX7qROKEiSxGI148lGN9ivl5P3
-1LTd2zWH8H9jrWvQmTHJp2q4OPdGFF4NjbnEwbJlLgzYzIs3amG9HVZ7tEFzGyMV
-X+panCi4/BXHTJ54pHF7TPeEJUoP4HnUC7X5vZ1tLik87GsX22b1USG2pypy46ky
-Lyx0CtqoY2vdd/Oygcwn2yqPhtQPyzdJlV+a+dK3Ii3eWJf5/DPl/H3gNzUqzJIe
-1yoDkXIQdrNdUw1aYOJLpkg1qgg5TjgVMA3elYeDgRvMZSZA3nHY+pvlYEC0FGpQ
-fCFuDY5FwiFPfdQ4MPRrMh1/Tla4ai0zO+5X0wIDAQABAoIBAGPeVVUVG3ZZysxf
-YPhTFfJmkPaHctAXoizN8XfJZ3318mP/wqjWcboxbQzHIOjb8saqzUlwiH9xgy1w
-xHxG7slbKIrAzpjv3VaMl9q2Ysrja8el1wFjez65Z7jUqji7JbuL8ymLZcmKFQYV
-1Q0FxNz9sb+ku9XxRDvtyJUbL07r7nB9ofHhRwUdkzjvFqlP7z9a39r/wP8n0AiO
-IMzjwTla+72ad9GGSut4BM3whRnt5/SlaHrgxFUiUZe7V2ONqiHkVv8OPcdJXIhr
-qA98ZOS4g+NjOWsJGd6nQqaQUbdoEPsSNeBvRSAyKPkjZRXxdH4M3zRE5xXIO3Rv
-2Vw1UiECgYEA7AdxXFOUwpntSKIvkf6HfHNFxqQ7Kt2igLD/oqWbhH9imHFQzIL2
-T2Bl3nfz8h/8SLNHnow/CHxY+AtDitqIM0eT5WayKvvj99n/kZ3ktR+W/ogqtmJ1
-qPVXoYe/MtnnhUTwN27yh5Qtwkag/f4HUCN+ciRGjLnvqbqeR8BVlKUCgYEA6imi
-uIJb3ZFzMqP/mZkveXLtebik6ewq6WL9X9fYxo0dfaJ0prnZOgwF09r23udmzgNs
-6TQKO0+t/1+KCnu4WGJP+r1UVlzhT2OZwWQeP6SkWI5HngN6rUKRz9ZpyKbCf3p/
-3aXzCIbCSir48/U5NeFV+ZW5zZIT4vjT4OBMeRcCgYEAjYitBoMsKkFpU2vGANLT
-ZeItzj7/J+Y9otQAhwStrBdDBUy1+dzyXIjerTfa9a8Odn3jDYTl6229YAhJGUOP
-9sgOPLqHTlU7z41OvoAi/CWUQs9eZX3HNmjggoFHOvQF7Bl5vpPuJTTK97uooIYq
-1nhiQ4nTMGHzACFi9n20E+0CgYA9MOAq4zXqcqXt59lVZCh5zkSseEnwLx0PnCSm
-NRua1ymLuTZOk2ZmutmTATdeDI5a6548WrMJyMqpX2gNUUNdJWPNHEi7wM3IFzkI
-xob5Dqc71QJ2EmaufCTqcRd3rcxJz9M2MrWN1tlS1GP2LHDk2ZvD+xJw2+mCeI+1
-xid1xQKBgQCfsF4s9hohZmucz6A2DK8n3cPOpfykc+rnV4IMcRXj0S9yJbina5JE
-c/yLNHWSf7I4KSvmecfen7yLL7q0mNNzWQLwjZcPHLyJvRsLEo6prYP8f/AVGHYT
-jSaFhWM3D2/fh+Bt+eCLE2C/NOF5q8AL792Fi/d/wTdC3ZQLajGMAg==
------END RSA PRIVATE KEY-----
