--- tests/backend/mongodb-ssl/certs/ca-chain.cert.pem
@@ -1,72 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIGQDCCBCigAwIBAgICEAAwDQYJKoZIhvcNAQELBQAwgbExCzAJBgNVBAYTAlVW
-MRQwEgYDVQQIDAtDb3NtaWNzdGF0ZTERMA8GA1UEBwwIemVyb25lc3QxLjAsBgNV
-BAoMJURlY2VudHJhbGl6ZWQgTm9tYWRpYyBBbnRzIENvbGxlY3RpdmUxFjAUBgNV
-BAsMDUxlYWRlckxlc3NMYWIxDzANBgNVBAMMBkROQUNMMzEgMB4GCSqGSIb3DQEJ
-ARYRZW1wdHlAZG5hYy1sMy5pb24wHhcNMTcxMjIwMjI1NjI2WhcNMjcxMjE4MjI1
-NjI2WjCBqzELMAkGA1UEBhMCVVYxFDASBgNVBAgMC0Nvc21pY3N0YXRlMS4wLAYD
-VQQKDCVEZWNlbnRyYWxpemVkIE5vbWFkaWMgQW50cyBDb2xsZWN0aXZlMRYwFAYD
-VQQLDA1MZWFkZXJMZXNzTGFiMRwwGgYDVQQDDBNJbnRlcm1lZGlhdGUgRE5BQ0wz
-MSAwHgYJKoZIhvcNAQkBFhFlbXB0eUBkbmFjLWwzLmlvbjCCAiIwDQYJKoZIhvcN
-AQEBBQADggIPADCCAgoCggIBALXhmoEkuqmmmCu87u7lrPMGz5fzs8FWDjzLEwOE
-uZeFCB8qvdasDxcTZzOSYXOGv4dI4w7QjM+5RRHVcZimGzchFxEyfxmYxLSIYWbG
-nP4NxI0ZeN3GNW2WE98owVhIvjK4fOOwc6VYit0RWgv7Q930Fi/Z2Yi1kXdxPxOi
-xXGIt8WH7GFlpgbL8pVLp5b7j+M3PP+fFI6V4eEo3v3cq9bCFjRCbPak6PvJZXEL
-uOTAGtiiFvpqQnKRE3Hfk0YUz7Mkwi6ojCfuUJ3HZztSdkeaCOT1qDuUOQzTGmlu
-A+ZTMEULPRx86Rt3B+p/0cwGcj5kv/AdbxMKme1J/vdaqPkxudcXdgNdr6rzKUb9
-kDkFoIxSa2hzVQhdVWTql9H7nOn9tYxYhIgxG1cEvGX/JTJlGvGj61cs2ZIYfcZe
-JDIzklzR2t1Z0b1ObzJApPOjhIgHNGx4mTwvGt7USsBknmAaValuF1Bd0cIYHQbx
-TGso9VjcPCAYX9WqBCm+zx5cHJAVeSgKrEds+5VQk8Sq1Ue7vIz0XaautvcYJ2ri
-SvLDmKvnB7Uql2zBzaNT0JUunfxgoSH+VWOGYt/iGwq2fv7gwM/1qDw+lWsRKYcP
-tfuJzfA9i3PWOAMt6gfzNeAG+CuBP0eniNsbfSZEFSqGQksp9d70oHAi+jypxeTo
-kV19AgMBAAGjZjBkMB0GA1UdDgQWBBQCF8NWbOOWvQ4kSeE0KVw/SNrTBjAfBgNV
-HSMEGDAWgBTwBHAMCWmWUOt4fwVJ/e9B5Vw//zASBgNVHRMBAf8ECDAGAQH/AgEA
-MA4GA1UdDwEB/wQEAwIBhjANBgkqhkiG9w0BAQsFAAOCAgEAYmglpKWb0/Cd0lmk
-D8VgfBsuc5x84z0vVoPjbllLOqfT38ymHh37IFESb4osCX+uaKSs/KMNjSSzKGg+
-GeYzf0iR2yGjezwTFDsIORjuPeoaMZjSVgxzenoIVqd8ptjBevkPzPXAlwkrQ5cl
-Hsz3DXmBjkQ9dLpPYcqADr4ir8XnQFMFsXvX3rpsBZ5E6J7HeU98BHoPm4K0NrY+
-+7P3nhlfyXrp5sSDUmkl2urZmlO8lQrIoGhz7vP/u9AER4YAvNG7WNDOzkMRwa3j
-e05n2RCzaoev8q9SEtF5ZoIS3Kr9HjGbZ8m9rRMeKq8BpT0CYwR8aYM8CAtaqkeH
-6/kSnU+7igOmNyaDamB5DvZ/cqNRLVqET03POe+n8HXPWT6D2NcCzkC6WAfoDJi5
-6ktLszEFMCKSXQyRz80J05L05iGyVwk4PHqH7eVKPFvU14CC00y/v5d3Is/jQTKw
-PZ6ma2tQ0Za9YDRQNANhXIq27XH9Hdyh56sD9LU7GoVG1+R1B4c/AL+mBMBZW1dG
-iPWI4BSLQe8RR3L+ioXdN3d9CkJYu6V6DEL36GiH48wfar7bJN2+Dl6Lc+wgnomw
-SHECBrOlJ+sjnSeqTFZK8fxs3QymRMQftm2CzXC2qVlOCILjnVrzNLo9ZUF+1Ljy
-JFIv4reqq/pBswBULIRBV/mw0/U=
------END CERTIFICATE-----
------BEGIN CERTIFICATE-----
-MIIGSjCCBDKgAwIBAgIJAI/Y8AsPkyadMA0GCSqGSIb3DQEBCwUAMIGxMQswCQYD
-VQQGEwJVVjEUMBIGA1UECAwLQ29zbWljc3RhdGUxETAPBgNVBAcMCHplcm9uZXN0
-MS4wLAYDVQQKDCVEZWNlbnRyYWxpemVkIE5vbWFkaWMgQW50cyBDb2xsZWN0aXZl
-MRYwFAYDVQQLDA1MZWFkZXJMZXNzTGFiMQ8wDQYDVQQDDAZETkFDTDMxIDAeBgkq
-hkiG9w0BCQEWEWVtcHR5QGRuYWMtbDMuaW9uMB4XDTE3MTIyMDIyMjEyM1oXDTM3
-MTIxNTIyMjEyM1owgbExCzAJBgNVBAYTAlVWMRQwEgYDVQQIDAtDb3NtaWNzdGF0
-ZTERMA8GA1UEBwwIemVyb25lc3QxLjAsBgNVBAoMJURlY2VudHJhbGl6ZWQgTm9t
-YWRpYyBBbnRzIENvbGxlY3RpdmUxFjAUBgNVBAsMDUxlYWRlckxlc3NMYWIxDzAN
-BgNVBAMMBkROQUNMMzEgMB4GCSqGSIb3DQEJARYRZW1wdHlAZG5hYy1sMy5pb24w
-ggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIKAoICAQDTwGwumwajfqm7N0LKUR4I
-zxUL/qdCxOX0TLLGCQ9RppgJVudx0myggYKMupboI0Yf0z8pSsIQauPoLobAeoOg
-71Z24iCiVMM7pJitoZxeMTu5ckWP7lDXeFUEt8cIpfmweSfariT2CK3mTgcJXxGk
-qv4x4VPpzvT9CdXSYd7Vajl3lclc9RBLy+1iLO88MkcnSmp3/zCujPuBtEfXBzVt
-AHgBUojebcEADyT9t33nXuVPUlUfT29LGi9UmohAsPSthj/if8gFfz2M/HXftyJv
-qe/TxAcKpJviCOVf+BD9/pcjYREYQyO3V6SzGXi/D6ZlpQMsus27y69gZ8bZ4zKa
-q7eneZb8Lrlu4Qb9+sdQORvzB5uJ+lshH6Vwyvkl1ejwsGT8HjqQX5T3UvshFg8n
-28CwpvRNDWh7AA/i/+0w4mf6EXfute2M/mQbWPB69fzwv2ReXFTfUZAoT/mq/toU
-yrsjw5a1saUlDxqkZn5CNuWS1ERnHbpuYTh4lvuMeJxkCBV9dZCI4PGGHaGAAiGV
-PN+1YnYggPpHoeCY6iOyTrKkLEDChEofEvM1Z96W1TFgtyx5vUDYsgtpj/HS5BL+
-7ZtjqsS8fXfBO39dgPQ1fd+vQkVeBFi7wvWYgxMpPP3GSP+iwNGDNcZnBzrKpHAt
-FYp5AVM35D0oCU2W4sf6zwIDAQABo2MwYTAdBgNVHQ4EFgQU8ARwDAlpllDreH8F
-Sf3vQeVcP/8wHwYDVR0jBBgwFoAU8ARwDAlpllDreH8FSf3vQeVcP/8wDwYDVR0T
-AQH/BAUwAwEB/zAOBgNVHQ8BAf8EBAMCAYYwDQYJKoZIhvcNAQELBQADggIBAGcw
-isfwKZ4/uhR7WKrViOlhOqcl6ju/f6tZvzM1rJ4OuyKnD4UUKXIAprv2OcjJCkEt
-WkOPIIACQkqdbzm5brBJMBKD4PSz7F4gq6h2npw/7XZGVDQbWVRmS9gkS1v5qIh+
-bckbyuNcA45o19i6P2JhxYMhQWx7hrv2DDsSEPy43QcdmAa1dQCVtjLb1cnMCvvE
-D0gRljyeQwoZBQTFYHTrIBLmZ5s2+eTvor+x+PAK3JX711ldhIOuecGVOXmbT4Ep
-dvDQQdXF0fQj0CDXcJk9YYXLy8uOGKaPql4mp+s5sNZ6U1DocZGmYyBhDbeOrW6C
-rAFVKTx6akdOdr6NqMTV4ONeaZpNWpYiUetiqYM3FTajVpIomGIo80ulsyKvPZ+7
-SR0/DK9QxvbFUskpcPSaxzXUhdLmzyGMjHCDcfEKqVq7NNQTfrTQfkTlSRa5XlMK
-5DxYh3Tf2C+I1M2YxCIES3hzNB1ZfkJyTFK/ofsqqXI2BF2VreZ13QSWXQeabMho
-iPUINNkwydk8NdCX6nDLlcLNxa/0Xlod3Xpxx2mRr7qS6Mv2uLQJfTsEj2fz4ks3
-Bpuk2ySGYtpZQ49iKMG0q/+6Fc4vtrZRy0EgZkCqpCgLey402uk1MthwG1RotaCa
-FzSxRdqTYrhc9DrDTLDnkVK36UBkIq1GFStnFMgE
------END CERTIFICATE-----
