--- tests/backend/mongodb-ssl/certs/crl.pem
@@ -1,19 +1,0 @@
------BEGIN X509 CRL-----
-MIIDKDCCARACAQEwDQYJKoZIhvcNAQELBQAwgasxCzAJBgNVBAYTAlVWMRQwEgYD
-VQQIDAtDb3NtaWNzdGF0ZTEuMCwGA1UECgwlRGVjZW50cmFsaXplZCBOb21hZGlj
-IEFudHMgQ29sbGVjdGl2ZTEWMBQGA1UECwwNTGVhZGVyTGVzc0xhYjEcMBoGA1UE
-AwwTSW50ZXJtZWRpYXRlIEROQUNMMzEgMB4GCSqGSIb3DQEJARYRZW1wdHlAZG5h
-Yy1sMy5pb24XDTE3MTIyMDIzNTgxMFoXDTE4MDExOTIzNTgxMFqgMDAuMB8GA1Ud
-IwQYMBaAFAIXw1Zs45a9DiRJ4TQpXD9I2tMGMAsGA1UdFAQEAgIQADANBgkqhkiG
-9w0BAQsFAAOCAgEASR0hqeSHzjupRqnV23EA6SVFpBm6iRJBhAPqN5ZjUtrBGtwT
-8ID8T3uMEUNLmw+BbIEEyJwjD4ZuGLoE3nImVEy+b9xoAoyD0OFzmWsLmB9mL31X
-hS1Wy9x05Ui7Q0mTVcq5wP5tK86CT8uNLqKbBVPnNJggDkRBGmrw6bJaZIewgw5i
-FHvvpefNBNFpZs/LzGtgMsV47TOX5QQ+R10bNcCDgwZ0PlyQKGP0suo8JIJgnX65
-gS35dJk3uAKNY949tQX1p0iCfOvqv6VQmEN/Xh1mPleejKhA2tMmI106lsBReHcQ
-ftuTGX3GWCLStIQ62vDto+2cr9001HeXBvgp4EupVQmLJ4/RgdmgQ0w0yxFuzbtG
-p+Ac1Mf8bUFhT0rMyvqYAkS8cZmbDQe0wt4ZC1Da3bYabAcODKfx+Rtx50Gg3oLD
-2xdZ3VD0LLbti68yDRG+0nJj4uPy2db3FP1AFp6Imy/InMvl2L4CC7V8H6tNAHeP
-z/SVSbv95+Zz60hZNheAChDL/VZzWi1RNBy01F9Txv8t+lTc1Cit4A63Z/bIPaju
-XWq9js3qGhEWyyTIoNEqcUmlipDbHTTPyBpzP/RFp5Q7NrAXWfh9BLtx+u26GKu9
-Bi9POgo8g+/gRjl/r3aNMvw5nd5OdyBFoANqFQNlIvaHObd8wYWI6JpeU0w=
------END X509 CRL-----
