--- tests/backend/mongodb-ssl/certs/local-mongo.pem
@@ -1,63 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEAuD40mDMjJDc1jVggh2fNn7b0q+Hwvpm1ZOWmlDwZrud5g6YY
-JZQ/MfkZcGoFzTO8ZhF8n2Rw8ZQRcCB5co513Ljw3o2yDXTrtF355kB0EaD9zXPg
-t9yLliU1Mm6ovJMII/U1uCGTjlhfIRpPmuNSFPqfsstEii8IuICVqISb2DTDjfZH
-lUd4v3HQVsSQARz/g3K/vWOXQQv5p8pHDfz75wdMlEmSO5H49Lo423HwTxY14ZRt
-Oc+yCc5LXNEti6aX0xI1pKrLK6OcT3bdiyAT755k31PpzwhaYm0C4/09TJR0SGqG
-d199FJxtZtn1XOczks2X4I6FCtI1mOZjsIISDQIDAQABAoIBAHUZ5GfQKMe952Wc
-25/0U/btW/lbM56RJ3+njwasZE6MhOktyi8oWDo07vxKBsHz3gyplKHiPv7sc2Dm
-4wnyjerqCP4K5V40MBOFn22/DYae4nGQ8wUpQfb+Ux9owgqJESV5ltkfTh+s/1mU
-1X/Gd8cpvBLxqvCBEHPkFjrzffIeEphlPjUi+hfN9tlNtkSm7hBMYD9eY5kLZDNN
-eLicCXLgNQgf2cIyITp91hZTiWJORZyK7WxVnoOKYG7wASgniGN5cLxu205JIMpo
-JeWFHtvOZitMUhhXSq+VZE9lycq9Y6lObhnb/8CrOd65o5EXg6CSHFGxa4p93xn7
-BqT2RIECgYEA5lFo5R4TJpWbxs58v/HgL5JpRg5KoV11XV+jZLCE7LjjtYKDhobZ
-tgG5ldEs9bpG+DEp/LpJgJnsHL75kQb1WFD/W5azm2Mqk4D0s/Dk4xj9H3RnNK08
-JSKD5e4/8EFqPG/OG4ekqF/MRVB1Ag/uEE2t4IGkGazvlNlCOFXJiVECgYEAzMmN
-QgzMwWk/sWe0N1hBYWHSwu6Xu0Xw0vzz++gm6K1Jb9tT/c2/FkPTjzy0auEq/KxR
-vtzMvhJsib1LSdGMYb8m/yO2T/4ixtPWWMSVEMzpaVko+kgBEYZEnQwY9yVNW5YT
-RSmDPNzEfQk1VJXU+71ArWD4U3jehmVh3SeITf0CgYEA1CC9FPFrQAlhbW03Nu6G
-xin6bfyxvge/Fh4E9rMh6omrqLypeRjwohLapGMBHt3Pib52ZoAJPLoH3r4CTBH1
-nL8VQnz9tatTRoHqX2UK2yNj1zm16K15jNZKV1blatpFB58OXQrEnBOHp3ugGSMk
-570z3WhqBP/jtduYgxx4P9ECgYA1mIpfvvnXpLQGuh1wXqi909xVKByN5sAgeBNi
-2l9UhgWyQLS6uB5KtXWpBzowvr2BY5gBW8g59phxdBlAOJeE/YI3RCFyzhBL+SUg
-FlLtErQD3lHbLv0YmQtcDs554EleruhcMYEnbKAAGWjbulsDh4Rl2/sh3YBfrCjw
-uGf0AQKBgQC4gUBKk5s9KLOtc9yValKHvU04P1wQWaoBPR6dTCrw/YAktqR24Zxl
-z4V6pfvLkVWPVAZ2dZxerVG0hulP+tvwW2e76RNgFFlnNN0RRz07qWEwkKgJrEYW
-1s95T7W0IigNugIUb2rUJ0Ld5Taa7C4z21IDU8lvO2JYBYFEZ9RaaQ==
------END RSA PRIVATE KEY-----
------BEGIN CERTIFICATE-----
-MIIGXDCCBESgAwIBAgICEAMwDQYJKoZIhvcNAQELBQAwgasxCzAJBgNVBAYTAlVW
-MRQwEgYDVQQIDAtDb3NtaWNzdGF0ZTEuMCwGA1UECgwlRGVjZW50cmFsaXplZCBO
-b21hZGljIEFudHMgQ29sbGVjdGl2ZTEWMBQGA1UECwwNTGVhZGVyTGVzc0xhYjEc
-MBoGA1UEAwwTSW50ZXJtZWRpYXRlIEROQUNMMzEgMB4GCSqGSIb3DQEJARYRZW1w
-dHlAZG5hYy1sMy5pb24wHhcNMTcxMjIxMDY0MjU4WhcNMTgxMjMxMDY0MjU4WjCB
-tDELMAkGA1UEBhMCVVYxFDASBgNVBAgMC0Nvc21pY3N0YXRlMREwDwYDVQQHDAh6
-ZXJvbmVzdDEuMCwGA1UECgwlRGVjZW50cmFsaXplZCBOb21hZGljIEFudHMgQ29s
-bGVjdGl2ZTEWMBQGA1UECwwNTGVhZGVyTGVzc0xhYjESMBAGA1UEAwwJbG9jYWxo
-b3N0MSAwHgYJKoZIhvcNAQkBFhFlbXB0eUBkbmFjLWwzLmlvbjCCASIwDQYJKoZI
-hvcNAQEBBQADggEPADCCAQoCggEBALg+NJgzIyQ3NY1YIIdnzZ+29Kvh8L6ZtWTl
-ppQ8Ga7neYOmGCWUPzH5GXBqBc0zvGYRfJ9kcPGUEXAgeXKOddy48N6Nsg1067Rd
-+eZAdBGg/c1z4Lfci5YlNTJuqLyTCCP1Nbghk45YXyEaT5rjUhT6n7LLRIovCLiA
-laiEm9g0w432R5VHeL9x0FbEkAEc/4Nyv71jl0EL+afKRw38++cHTJRJkjuR+PS6
-ONtx8E8WNeGUbTnPsgnOS1zRLYuml9MSNaSqyyujnE923YsgE++eZN9T6c8IWmJt
-AuP9PUyUdEhqhndffRScbWbZ9VznM5LNl+COhQrSNZjmY7CCEg0CAwEAAaOCAX0w
-ggF5MAkGA1UdEwQCMAAwEQYJYIZIAYb4QgEBBAQDAgZAMDMGCWCGSAGG+EIBDQQm
-FiRPcGVuU1NMIEdlbmVyYXRlZCBTZXJ2ZXIgQ2VydGlmaWNhdGUwHQYDVR0OBBYE
-FKshZxJ5Md78vhJwU4B+l+yhBAM0MIHfBgNVHSMEgdcwgdSAFAIXw1Zs45a9DiRJ
-4TQpXD9I2tMGoYG3pIG0MIGxMQswCQYDVQQGEwJVVjEUMBIGA1UECAwLQ29zbWlj
-c3RhdGUxETAPBgNVBAcMCHplcm9uZXN0MS4wLAYDVQQKDCVEZWNlbnRyYWxpemVk
-IE5vbWFkaWMgQW50cyBDb2xsZWN0aXZlMRYwFAYDVQQLDA1MZWFkZXJMZXNzTGFi
-MQ8wDQYDVQQDDAZETkFDTDMxIDAeBgkqhkiG9w0BCQEWEWVtcHR5QGRuYWMtbDMu
-aW9uggIQADAOBgNVHQ8BAf8EBAMCBaAwEwYDVR0lBAwwCgYIKwYBBQUHAwEwDQYJ
-KoZIhvcNAQELBQADggIBABafbpGaUsKK5ydJeotIvF4Rf6/KHo0kBuhOgdvpXaVx
-Epz2QFzMiHSa2WXFyc9oLsX7VecbFVSCzDmXYzF2wJ3zRmrQqfgbbrKMuZf3Isqp
-XdtiYrdZvPJlPDaB8zWsmR/G9guumZicbhMR5OyrDJjYP4mHiBAnSsJpxO1ZaJHo
-LBOjzLXeCOHRCcruMztOPS0+dGL/QL8hZpxrfBXWnDEBCQ5fre63j8aKKy4jwpjF
-akjI/Q62RcUvfhMhBm1jFYhnycnPz7z76tEvm10eQpNuRJZfreY5GzQSPpMd8r8w
-z9OC49HL5I7TpDW69Il6Ie+tEtbkyZ2411orKirTq1Yi15cmUHQZ/VHLPpVRorKJ
-tUxH/otYCebB/W1afX1Y5rCAgC9/NCTlfN2oV08/TgP4jam0p3xLyLPG2qQa0Aza
-96IbJdIH9mJkxK6BujFXiI1LNVgskFyYDlseXGUvn4Uwv4+uW/9HUNP39n8BVMwv
-vqabJzoqTkcjQ00PmzcHu+PtwyImDqnD1oPL4yj32yEXN3LmKNQ7mG2YU+4kxg+e
-ll5oxBu0Rtblp/fLo0UtuuCEarJglVduAA2ARLktzzV5ekIZDshUMttkY4HbcUMA
-KJG8+Ci0PAhgLBd9LjagdBtqUIv4DK9+ZuoHLLyGNaqcOnmm/tRqdacUnKTPhhce
------END CERTIFICATE-----
