--- tests/backend/mongodb-ssl/certs/mongodb.pem
@@ -1,63 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpQIBAAKCAQEAve+E6Vsf4i7YNnqn60NBihi5wFG44KGb+zeU1W45Fzr2wL4l
-aQHXie7wsvdJrtw0j19Qb+mUTIQ9qaysZA04rZEYZLj6zrMC5zy41OfqMhO6vXZ6
-f6SAL0Z+BfIroNj/hukcdBzMkXzZcrL/WG5Tr9SKfUY3ue4a4HW9PhHHp7cfB2AK
-j/YxKioHIXQeKffv8q5EJ+HuHDOMFl56jubPiThDsDX90fzDUKsYTr7AobGVYCk6
-sqQ+tCiOS98KZjQvmn9ufAReogAA4aKty79aM8j5PCbhykrcneTJG4Yi1Luxf20B
-Z1rtO/g1T5XqxCAlHLSv1vTc1id0Yhna4HqRzQIDAQABAoIBAA87jmVIsZ3qM9zq
-qVzs3+y7XupVo4Ygb+oN4xna5Sk/yVpb5ZQuWaLuTs+LlZU4g96bvk+7Y/FfCbL5
-7jl39fu1j0dErpi1Mp5o+Ena76Vkf62COwDRqJYiCSlg6q7TuGqRd364cI+ae7X/
-NQtajcukOBcj8oQu3Gj0CGFZ+7EvkxFV4rioRHKbwV16MEXMl6DyWzkMtlYVn89k
-vX/rvuvtLvxY5FL4zHIarm9hBf0QFPq2lqfunCPWv+UM6rzFAYsqPC0evio00TEN
-Sgq7B9L6tJGj+ZPjDfqo5+OG43Y/VSnybhM9FSW8pkV7zRTqnvudLPsyPWHHAxKJ
-5NxUvwECgYEA7EKfIkBBMaFDGg3LeN7xzj569eQz5lTwZcBXNwd2O8I9A7Zs7fnE
-3dGz8aIcoZToS3BJX/mQULNmhEqSQgNVYNNY1+ezFiZ1iohA+tiLfa0ut+G1QWfU
-B2RdqkIK6VaRqFc8ct9QYHQ4lwTNwBqjJL5M7LUEHTOKMAs3JAC461UCgYEAzc4P
-B43nyy5+Tm3ibGee/+yaXCA5jhJFlxcStwcvLCV3aAj82kY9YUNkOg52OBSsWvIJ
-3rEHQzeO0CbckuwKJBpaDdf5c6//lNh7Q+YY2eUqDuB0+gLGOwjgOei8vCpA2+Uw
-jkFcW4wLZ6MhqUNqG7rwgBg8s8ADgjfcgluWPJkCgYEAtt3iGjqtu2Jcc8AVq5q0
-grRtrAFUAoXgBIHZf6hw5VQ0sbjzm0bmD+MOaorVqmN9YVaqFHUwNKaHRQILLrkx
-zl4HuCWIwi5GcFGCKiaN4Vq4fG9hQhzBSm9u6CdyFqWA3FGZYjhjGCUc4BdBwd1c
-RUZeeCCr7IZGAWXM6jr85nkCgYEAlx0bSge3UChGU945aDG5vN8FBGog+KGP+X3W
-nw+mYnrsLzaSOs9iWF18ZWijA1CisXCRMLO5ynAr68jtIvAwDgxPXWTc5EYxmyBN
-ejCkxLGQ7TWAjuakmUpOIhczJdK8pvmlan6xiwLMiP9uQvOeINhlqRii0bVGnkHY
-JtuK99kCgYEAiOU8CUf1XeucHXqvjZ6G2J0OQLr9fr8grL8o1GOekjWw6rrWDndR
-aeE17Eb6coXgKcKBKuUM0hRcaWdlfzU2U97ZGaynGVW5kAUSwK1huKBeiwUsqhhD
-Ur/cQ5+7852Dkt3+3PzkMEZiXUbZLzQs/VHh1G3c+aiDb1ff/jKttqU=
------END RSA PRIVATE KEY-----
------BEGIN CERTIFICATE-----
-MIIGVjCCBD6gAwIBAgICEAIwDQYJKoZIhvcNAQELBQAwgasxCzAJBgNVBAYTAlVW
-MRQwEgYDVQQIDAtDb3NtaWNzdGF0ZTEuMCwGA1UECgwlRGVjZW50cmFsaXplZCBO
-b21hZGljIEFudHMgQ29sbGVjdGl2ZTEWMBQGA1UECwwNTGVhZGVyTGVzc0xhYjEc
-MBoGA1UEAwwTSW50ZXJtZWRpYXRlIEROQUNMMzEgMB4GCSqGSIb3DQEJARYRZW1w
-dHlAZG5hYy1sMy5pb24wHhcNMTcxMjIxMDQzNzQ5WhcNMTgxMjMxMDQzNzQ5WjCB
-rjELMAkGA1UEBhMCVVYxFDASBgNVBAgMC0Nvc21pY3N0YXRlMREwDwYDVQQHDAh6
-ZXJvbmVzdDEuMCwGA1UECgwlRGVjZW50cmFsaXplZCBOb21hZGljIEFudHMgQ29s
-bGVjdGl2ZTEWMBQGA1UECwwNTGVhZGVyTGVzc0xhYjEMMAoGA1UEAwwDbWRiMSAw
-HgYJKoZIhvcNAQkBFhFlbXB0eUBkbmFjLWwzLmlvbjCCASIwDQYJKoZIhvcNAQEB
-BQADggEPADCCAQoCggEBAL3vhOlbH+Iu2DZ6p+tDQYoYucBRuOChm/s3lNVuORc6
-9sC+JWkB14nu8LL3Sa7cNI9fUG/plEyEPamsrGQNOK2RGGS4+s6zAuc8uNTn6jIT
-ur12en+kgC9GfgXyK6DY/4bpHHQczJF82XKy/1huU6/Uin1GN7nuGuB1vT4Rx6e3
-HwdgCo/2MSoqByF0Hin37/KuRCfh7hwzjBZeeo7mz4k4Q7A1/dH8w1CrGE6+wKGx
-lWApOrKkPrQojkvfCmY0L5p/bnwEXqIAAOGircu/WjPI+Twm4cpK3J3kyRuGItS7
-sX9tAWda7Tv4NU+V6sQgJRy0r9b03NYndGIZ2uB6kc0CAwEAAaOCAX0wggF5MAkG
-A1UdEwQCMAAwEQYJYIZIAYb4QgEBBAQDAgZAMDMGCWCGSAGG+EIBDQQmFiRPcGVu
-U1NMIEdlbmVyYXRlZCBTZXJ2ZXIgQ2VydGlmaWNhdGUwHQYDVR0OBBYEFM+nlfQo
-sgAM8pg4f990+am3gNKzMIHfBgNVHSMEgdcwgdSAFAIXw1Zs45a9DiRJ4TQpXD9I
-2tMGoYG3pIG0MIGxMQswCQYDVQQGEwJVVjEUMBIGA1UECAwLQ29zbWljc3RhdGUx
-ETAPBgNVBAcMCHplcm9uZXN0MS4wLAYDVQQKDCVEZWNlbnRyYWxpemVkIE5vbWFk
-aWMgQW50cyBDb2xsZWN0aXZlMRYwFAYDVQQLDA1MZWFkZXJMZXNzTGFiMQ8wDQYD
-VQQDDAZETkFDTDMxIDAeBgkqhkiG9w0BCQEWEWVtcHR5QGRuYWMtbDMuaW9uggIQ
-ADAOBgNVHQ8BAf8EBAMCBaAwEwYDVR0lBAwwCgYIKwYBBQUHAwEwDQYJKoZIhvcN
-AQELBQADggIBAJak6j+P+NoPsNR5uNxSZpVDmnE6uthhzYi7q1Svu62y4H8xt9sS
-0QCpTDPntdBErpeq2nysYm9u+d0PUBfHD3iqlgHrEqFnZbixAcrB0DfqUWnclALD
-GAyy/6R93wKfXAQrzAbXTmMDwfkw1s0BDVhejAMVOniGcNl13LY0g935EgC14x1y
-JADeB7hWcPQO4HTPUm1X1vAGwAlXFd4Q8vY/kWqCezc9/TZRuzK/oRThLA1GNegc
-BZpzbHF4wv9SG95SEsCDAUN6Z350ylVyuRfOd68r8nq20NSBybwYhJALinYHJPxY
-4vgcAykqWp2Aff1HaKat7ps6iJBic8n+SNOmQNs2LWCppTjNLDaZPE8TIMF74pix
-NUvR6OCF433cYIBXRPRk6J4U67grC+r/yUTieUOIrfAMoXrM9VhqErJW1yk/H4/O
-yAsVUS26it3R+FbCSg1Pa2Dt9dgt401Y2xaEQ4nEC6zQblwQHq9rM364Ji47uFM6
-gsVLJXysJ/i+vFcHHIufCOJml0+azDzfD3XVRFChenQwwJVmKgwCtNEgy3+EoKbx
-zkV4w4eagLQZsyCtl5Ed2iawMQVV9ECAsqKU+IUN93rGHpBMEBzNpN94YehTcYC4
-gfQmny5f0yXm7XE2exjFbkPXpqOpBajFl2Wmaz6yqc+0CNeRxRpj4ksX
------END CERTIFICATE-----
