+++ tests/backend/test_connection.py
@@ -1,6 +1,9 @@
 import pytest
 
 
+pytestmark = pytest.mark.tendermint
+
+
 def test_get_connection_raises_a_configuration_error(monkeypatch):
     from bigchaindb.common.exceptions import ConfigurationError
     from bigchaindb.backend import connect
