+++ tests/backend/test_utils.py
@@ -4,6 +4,9 @@ from types import ModuleType
 import pytest
 
 
+pytestmark = pytest.mark.tendermint
+
+
 @pytest.fixture
 def mock_module():
     return ModuleType('mock_module')
