+++ tests/conftest.py
@@ -36,12 +36,11 @@ def pytest_runtest_setup(item):
 def pytest_addoption(parser):
     from bigchaindb.backend.connection import BACKENDS
 
-    BACKENDS['mongodb-ssl'] = 'bigchaindb.backend.mongodb.connection.MongoDBConnection'
     backends = ', '.join(BACKENDS.keys())
     parser.addoption(
         '--database-backend',
         action='store',
-        default=os.environ.get('BIGCHAINDB_DATABASE_BACKEND', 'rethinkdb'),
+        default=os.environ.get('BIGCHAINDB_DATABASE_BACKEND', 'localmongodb'),
         help='Defines the backend to use (available: {})'.format(backends),
     )
 
@@ -50,7 +49,6 @@ def pytest_ignore_collect(path, config):
     from bigchaindb.backend.connection import BACKENDS
     path = str(path)
 
-    BACKENDS['mongodb-ssl'] = 'bigchaindb.backend.mongodb.connection.MongoDBConnection'
     supported_backends = BACKENDS.keys()
 
     if os.path.isdir(path):
@@ -63,20 +61,19 @@ def pytest_ignore_collect(path, config):
 def pytest_configure(config):
     config.addinivalue_line(
         'markers',
-        'bdb(): Mark the test as needing BigchainDB, i.e. a database with '
-        'the three tables: "backlog", "bigchain", "votes". BigchainDB will '
-        'be configured such that the database and tables are available for an '
-        'entire test session. For distributed tests, the database name will '
-        'be suffixed with the process identifier, e.g.: "bigchain_test_gw0", '
-        'to ensure that each process session has its own separate database.'
+        'bdb(): Mark the test as needing BigchainDB.'
+        'BigchainDB will be configured such that the database and tables are available for an '
+        'entire test session.'
+        'You need to run a backend (e.g. MongoDB) '
+        'prior to running tests with this marker. You should not need to restart the backend '
+        'in between tests runs since the test infrastructure flushes the backend upon session end.'
     )
     config.addinivalue_line(
         'markers',
-        'genesis(): Mark the test as needing a genesis block in place. The '
-        'prerequisite steps of configuration and database setup are taken '
-        'care of at session scope (if needed), prior to creating the genesis '
-        'block. The genesis block has function scope: it is destroyed after '
-        'each test function/method.'
+        'abci(): Mark the test as needing a running ABCI server in place. Use this marker'
+        'for tests that require a running Tendermint instance. Note that the test infrastructure'
+        'has no way to reset Tendermint data upon session end - you need to do it manually.'
+        'Setup performed by this marker includes the steps performed by the bdb marker.'
     )
 
 
@@ -87,12 +84,6 @@ def _bdb_marker(request):
 
 
 @pytest.fixture(autouse=True)
-def _genesis_marker(request):
-    if request.keywords.get('genesis', None):
-        request.getfixturevalue('_genesis')
-
-
-@pytest.fixture(autouse=True)
 def _restore_config(_configure_bigchaindb):
     from bigchaindb import config, config_utils
     config_before_test = copy.deepcopy(config)
@@ -100,29 +91,8 @@ def _restore_config(_configure_bigchaindb):
     config_utils.set_config(config_before_test)
 
 
-@pytest.fixture
-def _restore_dbs(request):
-    from bigchaindb.backend import connect, schema
-    from bigchaindb.common.exceptions import DatabaseDoesNotExist
-    from .utils import list_dbs
-    conn = connect()
-    dbs_before_test = list_dbs(conn)
-    yield
-    dbs_after_test = list_dbs(conn)
-    dbs_to_delete = (
-        db for db in set(dbs_after_test) - set(dbs_before_test)
-        if TEST_DB_NAME not in db
-    )
-    print(dbs_to_delete)
-    for db in dbs_to_delete:
-        try:
-            schema.drop_database(conn, db)
-        except DatabaseDoesNotExist:
-            pass
-
-
 @pytest.fixture(scope='session')
-def _configure_bigchaindb(request, ssl_context):
+def _configure_bigchaindb(request):
     import bigchaindb
     from bigchaindb import config_utils
     test_db_name = TEST_DB_NAME
@@ -132,21 +102,6 @@ def _configure_bigchaindb(request, ssl_context):
         test_db_name = '{}_{}'.format(TEST_DB_NAME, xdist_suffix)
 
     backend = request.config.getoption('--database-backend')
-
-    if backend == 'mongodb-ssl':
-        bigchaindb._database_map[backend] = {
-            # we use mongodb as the backend for mongodb-ssl
-            'backend': 'mongodb',
-            'connection_timeout': 5000,
-            'max_tries': 3,
-            'ssl': True,
-            'ca_cert': ssl_context.ca,
-            'crlfile': ssl_context.crl,
-            'certfile': ssl_context.cert,
-            'keyfile': ssl_context.key,
-            'keyfile_passphrase': os.environ.get('BIGCHAINDB_DATABASE_KEYFILE_PASSPHRASE', None)
-        }
-        bigchaindb._database_map[backend].update(bigchaindb._base_database_mongodb)
 
     config = {
         'database': bigchaindb._database_map[backend],
@@ -193,32 +148,11 @@ def _setup_database(_configure_bigchaindb):
 def _bdb(_setup_database, _configure_bigchaindb):
     from bigchaindb import config
     from bigchaindb.backend import connect
-    from bigchaindb.backend.admin import get_config
-    from bigchaindb.backend.schema import TABLES
-    from .utils import flush_db, update_table_config
+    from .utils import flush_db
     conn = connect()
-    # TODO remove condition once the mongodb implementation is done
-    if config['database']['backend'] == 'rethinkdb':
-        table_configs_before = {
-            t: get_config(conn, table=t) for t in TABLES
-        }
     yield
     dbname = config['database']['name']
     flush_db(conn, dbname)
-    # TODO remove condition once the mongodb implementation is done
-    if config['database']['backend'] == 'rethinkdb':
-        for t, c in table_configs_before.items():
-            update_table_config(conn, t, **c)
-
-
-@pytest.fixture
-def _genesis(_bdb, genesis_block):
-    # TODO for precision's sake, delete the block once the test is done. The
-    # deletion is done indirectly via the teardown code of _bdb but explicit
-    # deletion of the block would make things clearer. E.g.:
-    # yield
-    # tests.utils.delete_genesis_block(conn, dbname)
-    pass
 
 
 # We need this function to avoid loading an existing
@@ -392,16 +326,10 @@ def structurally_valid_vote():
 
 
 @pytest.fixture
-def genesis_block(b):
-    return b.create_genesis_block()
-
-
-@pytest.fixture
-def inputs(user_pk, b, genesis_block):
+def inputs(user_pk, b):
     from bigchaindb.models import Transaction
 
     # create blocks with transactions for `USER` to spend
-    prev_block_id = genesis_block.id
     for block in range(4):
         transactions = [
             Transaction.create(
@@ -412,20 +340,14 @@ def inputs(user_pk, b, genesis_block):
             for _ in range(10)
         ]
         block = b.create_block(transactions)
-        b.write_block(block)
-
-        # vote the blocks valid, so that the inputs are valid
-        vote = b.vote(block.id, prev_block_id, True)
-        prev_block_id = block.id
-        b.write_vote(vote)
+        b.store_block(block.to_dict())
 
 
 @pytest.fixture
-def inputs_shared(user_pk, user2_pk, genesis_block):
+def inputs_shared(user_pk, user2_pk):
     from bigchaindb.models import Transaction
 
     # create blocks with transactions for `USER` to spend
-    prev_block_id = genesis_block.id
     for block in range(4):
         transactions = [
             Transaction.create(
@@ -436,12 +358,7 @@ def inputs_shared(user_pk, user2_pk, genesis_block):
             for _ in range(10)
         ]
         block = b.create_block(transactions)
-        b.write_block(block)
-
-        # vote the blocks valid, so that the inputs are valid
-        vote = b.vote(block.id, prev_block_id, True)
-        prev_block_id = block.id
-        b.write_vote(vote)
+        b.store_block(block.to_dict())
 
 
 @pytest.fixture
@@ -611,53 +528,6 @@ def abci_server():
     abci_proxy.terminate()
 
 
-@pytest.fixture(scope='session')
-def certs_dir():
-    return os.path.abspath('tests/backend/mongodb-ssl/certs')
-
-
-@pytest.fixture(scope='session')
-def ca_chain_cert(certs_dir):
-    return os.environ.get(
-        'BIGCHAINDB_DATABASE_CA_CERT',
-        os.path.join(certs_dir, 'ca-chain.cert.pem'))
-
-
-@pytest.fixture(scope='session')
-def ssl_crl(certs_dir):
-    return os.environ.get(
-        'BIGCHAINDB_DATABASE_CRLFILE',
-        os.path.join(certs_dir, 'crl.pem'))
-
-
-@pytest.fixture(scope='session')
-def ssl_cert(certs_dir):
-    return os.environ.get(
-        'BIGCHAINDB_DATABASE_CERTFILE',
-        os.path.join(certs_dir, 'bigchaindb.cert.pem'))
-
-
-@pytest.fixture(scope='session')
-def ssl_key(certs_dir):
-    return os.environ.get(
-        'BIGCHAINDB_DATABASE_KEYFILE',
-        os.path.join(certs_dir, 'bigchaindb.key.pem'))
-
-
-@pytest.fixture
-def mdb_ssl_pem_key(certs_dir):
-    return os.environ.get(
-        'BIGCHAINDB_MDB_PEM_KEY_TEST',
-        os.path.join(certs_dir, 'local-mongo.pem'))
-
-
-@pytest.fixture(scope='session')
-def ssl_context(ca_chain_cert, ssl_crl, ssl_cert, ssl_key):
-    SSLContext = namedtuple('SSLContext', ('ca', 'crl', 'cert', 'key'))
-    return SSLContext(
-        ca=ca_chain_cert, crl=ssl_crl, cert=ssl_cert, key=ssl_key)
-
-
 @pytest.fixture
 def wsserver_config():
     from bigchaindb import config
@@ -682,15 +552,6 @@ def wsserver_port(wsserver_config):
 @pytest.fixture
 def wsserver_base_url(wsserver_scheme, wsserver_host, wsserver_port):
     return '{}://{}:{}'.format(wsserver_scheme, wsserver_host, wsserver_port)
-
-
-@pytest.fixture
-def genesis_tx(b, user_pk):
-    from bigchaindb.models import Transaction
-    tx = Transaction.create([b.me], [([user_pk], 1)])
-    tx.operation = Transaction.GENESIS
-    genesis_tx = tx.sign([b.me_private])
-    return genesis_tx
 
 
 @pytest.fixture
